Fixes an issue where the video modal would be misplaced
