+++ src/components/ModalAnniversary/Footer/styles.postcss
