+++ src/components/ModalAnniversary/HistoryHeader/styles.postcss
