+++ src/components/ModalAnniversary/PowerBox/styles.postcss
