+++ src/components/ModalAnniversary/Slider/index.js
