+++ src/components/ModalAnniversary/Slider/styles.postcss
