+++ src/components/ModalAnniversary/index.js
@@ -1,9 +1,18 @@
 'use strict';
 
 import './styles.postcss';
-import Modal from '../Modal';
 import React from 'react';
-import Anniversary from '../Anniversary';
+import Modal from '../Modal';
+import CaresPackage from './CaresPackage';
+import PowerBox from './PowerBox';
+import CaresEvolution from './CaresEvolution';
+import CatalystWomen from './CatalystWomen';
+import BoxVideo from './BoxVideo';
+import Slider from './Slider';
+import HistoryHeader from './HistoryHeader';
+import Retooling from './Retooling';
+import Donation from './Donation';
+import Footer from './Footer';
 
 class ModalAnniversary extends Modal {
 
@@ -17,17 +26,28 @@ class ModalAnniversary extends Modal {
 
   getContent() {
     return (
-    	<Anniversary 
-        toggleMenuFn = { this.props.toggleMenuFn }
-        onClose = { this.props.onClose.bind(this) }
-      />
-    );
+      <section id="anniversary">
+        <HistoryHeader
+          onClose = {this.props.onClose.bind(this)}
+        />
+        <CaresPackage />
+        <PowerBox />
+        <Retooling />
+        <CaresEvolution />
+        <CatalystWomen />
+        <Slider />
+        <BoxVideo openModalVideo={this.props.openModalVideo} />
+        <Donation />
+        <Footer />
+      </section>
+    )
   }
 }
 
 ModalAnniversary.propTypes = {
   visible: React.PropTypes.bool.isRequired,
-  onClose: React.PropTypes.func.isRequired
+  onClose: React.PropTypes.func.isRequired,
+  openModalVideo: React.PropTypes.func
 };
 
 export default ModalAnniversary;
