+++ src/components/ModalAnniversary/styles.postcss
@@ -26,3 +26,23 @@
 	    }
 	  }
 }
+
+.l-header {
+  position: relative;
+  pointer-events: visible;
+}
+
+.viel:before {
+    content: "";
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    background-color: rgba(55, 53, 47, 0.5);
+}
+
+.background-image {
+  background-repeat: no-repeat;
+  background-size: cover;
+}
