+++ src/components/ModalVideo/styles.postcss
@@ -1,9 +1,10 @@
-@import'../../../styles/settings.postcss';
+@import'../../styles/settings.postcss';
 
 
 
 .modal-video {
 	position: fixed;
+	z-index: 50;
 
 	.content {
 		max-height: 100% !important;
